requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
insightface>=0.7.3
faiss-cpu>=1.7.4
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
onnxruntime>=1.16.0
python-multipart>=0.0.6
websockets>=12.0
tqdm>=4.65.0
pytest>=7.0.0
